# docs/swagger.yaml
basePath: /api/
definitions:
  domain.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  domain.ErrorMessage:
    properties:
      error:
        type: boolean
      message:
        type: string
      statusCode:
        type: integer
    type: object
  domain.LoginInput:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  domain.RegisterInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
  domain.ResponseAuth:
    properties:
      jwt: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
  domain.UpdateCategoryRequest:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: REST API for FiberInventory
  title: FiberInventory
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username and password to get a JWT token
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ResponseAuth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      summary: Login to the application
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create User
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      summary: Register to the application
      tags:
      - Auth
  /category:
    get:
      description: Retrieve the results for each category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get category results
      tags:
      - Category
  /category/{id}:
    delete:
      description: Delete a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Category
    get:
      description: Retrieve the result for a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get category result
      tags:
      - Category
    put:
      description: Update a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Category
  /category/create:
    post:
      consumes:
      - application/json
      description: Create Category
      parameters:
      - description: Category information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: createCategory to the application
      tags:
      - Category
...
securityDefinitions:
  BearerAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"